requirements.txt
# Core
anthropic>=0.84.0
openai>=2.24.0
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.12.0
python-dotenv>=1.0.0

# Storage
motor>=3.7.0          # async MongoDB driver

# Vector Search & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0

# Dev/Testing
pytest>=9.0.0
pytest-asyncio>=1.3.0
httpx>=0.28.0         # for testing FastAPI
